+++ lib/src/types/tool_call.mapper.dart
@@ -46,6 +46,13 @@ class ToolCallMapper extends ClassMapperBase<ToolCall> {
     key: r'canonical_path',
     opt: true,
   );
+  static String? _$serverName(ToolCall v) => v.serverName;
+  static const Field<ToolCall, String> _f$serverName = Field(
+    'serverName',
+    _$serverName,
+    key: r'server_name',
+    opt: true,
+  );
 
   @override
   final MappableFields<ToolCall> fields = const {
@@ -53,6 +60,7 @@ class ToolCallMapper extends ClassMapperBase<ToolCall> {
     #args: _f$args,
     #id: _f$id,
     #canonicalPath: _f$canonicalPath,
+    #serverName: _f$serverName,
   };
   @override
   final bool ignoreNull = true;
@@ -63,6 +71,7 @@ class ToolCallMapper extends ClassMapperBase<ToolCall> {
       args: data.dec(_f$args),
       id: data.dec(_f$id),
       canonicalPath: data.dec(_f$canonicalPath),
+      serverName: data.dec(_f$serverName),
     );
   }
 
@@ -124,12 +133,13 @@ extension ToolCallValueCopy<$R, $Out> on ObjectCopyWith<$R, ToolCall, $Out> {
 abstract class ToolCallCopyWith<$R, $In extends ToolCall, $Out>
     implements ClassCopyWith<$R, $In, $Out> {
   MapCopyWith<$R, String, dynamic, ObjectCopyWith<$R, dynamic, dynamic>?>
-  get args;
+      get args;
   $R call({
     String? name,
     Map<String, dynamic>? args,
     String? id,
     String? canonicalPath,
+    String? serverName,
   });
   ToolCallCopyWith<$R2, $In, $Out2> $chain<$R2, $Out2>(Then<$Out2, $R2> t);
 }
@@ -144,37 +154,42 @@ class _ToolCallCopyWithImpl<$R, $Out>
       ToolCallMapper.ensureInitialized();
   @override
   MapCopyWith<$R, String, dynamic, ObjectCopyWith<$R, dynamic, dynamic>?>
-  get args => MapCopyWith(
-    $value.args,
-    (v, t) => ObjectCopyWith(v, $identity, t),
-    (v) => call(args: v),
-  );
+      get args => MapCopyWith(
+            $value.args,
+            (v, t) => ObjectCopyWith(v, $identity, t),
+            (v) => call(args: v),
+          );
   @override
   $R call({
     String? name,
     Object? args = $none,
     Object? id = $none,
     Object? canonicalPath = $none,
-  }) => $apply(
-    FieldCopyWithData({
-      if (name != null) #name: name,
-      if (args != $none) #args: args,
-      if (id != $none) #id: id,
-      if (canonicalPath != $none) #canonicalPath: canonicalPath,
-    }),
-  );
+    Object? serverName = $none,
+  }) =>
+      $apply(
+        FieldCopyWithData({
+          if (name != null) #name: name,
+          if (args != $none) #args: args,
+          if (id != $none) #id: id,
+          if (canonicalPath != $none) #canonicalPath: canonicalPath,
+          if (serverName != $none) #serverName: serverName,
+        }),
+      );
   @override
   ToolCall $make(CopyWithData data) => ToolCall(
-    name: data.get(#name, or: $value.name),
-    args: data.get(#args, or: $value.args),
-    id: data.get(#id, or: $value.id),
-    canonicalPath: data.get(#canonicalPath, or: $value.canonicalPath),
-  );
+        name: data.get(#name, or: $value.name),
+        args: data.get(#args, or: $value.args),
+        id: data.get(#id, or: $value.id),
+        canonicalPath: data.get(#canonicalPath, or: $value.canonicalPath),
+        serverName: data.get(#serverName, or: $value.serverName),
+      );
 
   @override
   ToolCallCopyWith<$R2, ToolCall, $Out2> $chain<$R2, $Out2>(
     Then<$Out2, $R2> t,
-  ) => _ToolCallCopyWithImpl<$R2, $Out2>($value, $cast, t);
+  ) =>
+      _ToolCallCopyWithImpl<$R2, $Out2>($value, $cast, t);
 }
 
 class ToolResultMapper extends ClassMapperBase<ToolResult> {
@@ -321,27 +336,28 @@ class _ToolResultCopyWithImpl<$R, $Out>
     Object? result = $none,
     Object? error = $none,
     Object? exception = $none,
-  }) => $apply(
-    FieldCopyWithData({
-      if (name != null) #name: name,
-      if (id != $none) #id: id,
-      if (result != $none) #result: result,
-      if (error != $none) #error: error,
-      if (exception != $none) #exception: exception,
-    }),
-  );
+  }) =>
+      $apply(
+        FieldCopyWithData({
+          if (name != null) #name: name,
+          if (id != $none) #id: id,
+          if (result != $none) #result: result,
+          if (error != $none) #error: error,
+          if (exception != $none) #exception: exception,
+        }),
+      );
   @override
   ToolResult $make(CopyWithData data) => ToolResult(
-    name: data.get(#name, or: $value.name),
-    id: data.get(#id, or: $value.id),
-    result: data.get(#result, or: $value.result),
-    error: data.get(#error, or: $value.error),
-    exception: data.get(#exception, or: $value.exception),
-  );
+        name: data.get(#name, or: $value.name),
+        id: data.get(#id, or: $value.id),
+        result: data.get(#result, or: $value.result),
+        error: data.get(#error, or: $value.error),
+        exception: data.get(#exception, or: $value.exception),
+      );
 
   @override
   ToolResultCopyWith<$R2, ToolResult, $Out2> $chain<$R2, $Out2>(
     Then<$Out2, $R2> t,
-  ) => _ToolResultCopyWithImpl<$R2, $Out2>($value, $cast, t);
+  ) =>
+      _ToolResultCopyWithImpl<$R2, $Out2>($value, $cast, t);
 }
-